refactor: 카테고리 타입 정의 파일 위치 변경
- src/components/Aside/types/category.ts에서 src/types/category.d.ts로 위치 변경
